ci(test): Add CI workflow, test infrastructure, and path aliases
Establish a proper test layout separate from `src/`, wire up Turbo
`test`/`typecheck` tasks, add a GitHub Actions CI pipeline, and
introduce shared test support modules (`cloud`, `fusor`, `platform`,
`timing`) so test helpers aren't duplicated across suites.

Moves all colocated `*.test.ts` files under `test/core/` and
`test/providers/`, updates imports to use `@/` (src) and `@test/`
(test) path aliases configured in `tsconfig.json`, and documents the
test layout in CONTRIBUTING.md.
