+++ packages/spectrum-ts/package.json
@@ -67,7 +67,11 @@
     "dev": "tsup --watch",
     "generate:emoji": "bun scripts/generate-emoji.ts",
     "generate:manifest": "bun scripts/generate-manifest.ts",
-    "prepack": "bun scripts/prepare-package.ts --publish && tsup && node scripts/smoke-import.mjs && bun scripts/generate-manifest.ts && bun scripts/prepare-package.ts"
+    "prepack": "bun scripts/prepare-package.ts --publish && tsup && node scripts/smoke-import.mjs && bun scripts/generate-manifest.ts && bun scripts/prepare-package.ts",
+    "test": "bun test",
+    "test:watch": "bun test --watch",
+    "test:coverage": "bun test --coverage",
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "@photon-ai/advanced-imessage": "^0.11.0",
